+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Bugfixes:
 - Prevent interceptors from being able to modify `in` parameters (@stakx, #370)
+- Make default value replication of optional parameters more tolerant of default values that are represented in metadata with a mismatched type (@stakx, #371)
 
 Deprecations:
 - `Castle.DynamicProxy.Generators.Emitters.ArgumentsUtil.IsAnyByRef` (@stakx, #370)